pyproject.toml
[tool.poetry]
name = "real-life-rpg"
version = "0.1.0"
description = ""
authors = ["riisatoro <riisa.toro@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.10"
fastapi = {extras = ["uvicorn"], version = "^0.87.0"}
pymongo = "^4.3.3"
pre-commit = "^2.20.0"
PyJWT = "^2.6.0"
uvicorn = "^0.20.0"
bcrypt = "^4.0.1"

[tool.poetry.dev-dependencies]
ipdb = "^0.13.9"
ipython = "^8.6.0"

[tool.poetry.group.dev.dependencies]
isort = "^5.10.1"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
